+++ frontend/src/features/header-profile/style.css
