--- frontend/src/shared/constants/MAX_TWIT_MSG_LEN.ts
@@ -1,3 +1,0 @@
-const MAX_TWIT_MSG_LEN = 140;
-
-export default MAX_TWIT_MSG_LEN;
