+++ frontend/src/shared/constants/common.ts
@@ -1,0 +1,4 @@
+export const MAX_TWIT_MSG_LEN = 140 as const;
+
+export const defaultUserSrc =
+  'https://icon-library.com/images/twitter-bird-icon-png/twitter-bird-icon-png-18.jpg';
